+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.0.0-dev.1.8
+
+- removed entranceLabel, entranceLabelStyle, entranceLabelRadius from `Block` class, use arc openings instead.
+- introduced arc openings for blocks. They work exactly like openings
+  but are arcs.
+
+## 1.0.0-dev.1.7
+
+- more development on the API
+- update README
+
 ## 1.0.0-dev.1.6
 
 - breaking changes with API